# app/server/internal/transport/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vow API
  version: 0.1.0
  description: HTTP API for the Vow backend.
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Health
  - name: Auth
  - name: Users
paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service health status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '501':
          description: Endpoint scaffolded but not implemented yet.
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: Endpoint scaffolded but not implemented yet.
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh auth tokens
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refresh succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: Endpoint scaffolded but not implemented yet.
  /users/me:
    get:
      tags:
        - Users
      summary: Get the authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '501':
          description: Endpoint scaffolded but not implemented yet.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      properties:
        data:
          nullable: true
    ErrorEnvelope:
      type: object
      required:
        - success
        - statusCode
        - error
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: integer
          example: 401
        error:
          type: object
          required:
            - errorCode
            - message
            - userMessage
          properties:
            errorCode:
              type: string
              example: AUTH_001
            message:
              type: string
              example: Unauthorized access.
            userMessage:
              type: string
              example: You must be logged in to perform this action.
            requestId:
              type: string
              example: 6f0f4cc4bb9348e391b52aa72835ffdc
    HealthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              example: ok
            database:
              type: string
              enum:
                - ok
                - not_configured
              example: ok
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: correct-horse-battery-staple
        display_name:
          type: string
          example: Alex
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          format: password
          example: correct-horse-battery-staple
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: refresh-token-value
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            user:
              $ref: '#/components/schemas/User'
    UserResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          example: usr_123
        email:
          type: string
          format: email
          example: alex@example.com
        display_name:
          type: string
          example: Alex
        role:
          type: string
          example: user
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'